style : 메인 genre carousel 리팩토링 메인 페이지 genre carousel 컴포넌트로 따로 분리해서 리팩토링
